// tests/api/lib/utils/APIRoutes.ts
export enum APIRoutes {
    Users = '/users',
    Notes = '/notes'
}

export enum UserAPIRoutes {
    Register = '/users/register',
    Login = '/users/login',
    ForgotPassword = '/users/forgot-password',
    ResetPassword = '/users/reset-password',
    VerifyResetPassword = '/users/verify-reset-password-token',
    ChangePassword = '/users/change-password',
    Logout = '/users/logout',
    DeleteAccount = '/users/delete-account'
}